Ensure the self.filenames list is sorted internally.
This is important because even though the QListWidget UI is sorted
automatically, the list is not necessarily. Drag-drop events, depending
on the order in which items were highlighted before being dragged,
can result in a different or reversed order. This has implications for
popping the list later (e.g on delete events)